Add alg_inversion_number
