+++ crates/algolib/segtree2/Cargo.toml
@@ -17,4 +17,5 @@ package = "query_test"
 
 [dev-dependencies]
 test_vector2 = { path = "../../test_tools/test_vector2" }
+alg_inversion_number = { path = "../../algebra/alg_inversion_number" }
 assert-impl = "0.1.3"
